move hacky scrolling logic to its own hacky module
